moved ParticipantsDialog to classic header source implementation
Currently modules don't properly work with Qt as soon as u need Q_OBJECT all tested compilers will run into internal errors. It is not completely clear what the source of the Problem is and if it will be possible with (near) future Qt. 
But to be able to make progress we will go on with the classic approach to build Qt applications.