added TeamsCollector class and basic infrastructure for it
TeamCollector is used to filter all mentioned teams/crews while registering participants to make them available when planning a tournament.
To have an automated feedback, when registered Participants change (get added removed) a std:.function is currently used. In the future this should be replaced by a boost::signals2:signal, but the implementation is not working with the current implementation of msvc (probably due to C++20 feature missing).