second step to implementation of swiss style tournament
Now the result of the first round is taken into account.
Its only simply tested yet.
Still missing is that when the tournament has an odd number of contestants one will not have a match per round and this must not happen to the same contestant again in a tournament.
Or matches with three contestants must be enabled.
Another point that is missing currently is prevention of already happened matches.