fixed small bugs and unintended behaviour in matchmaking
The choosen apporoach is flawed when creating a lot of rounds, because it doesn't strongly assures that not every contestant skips a round, in case an odd number of contestants are participating.
To fix this a more overviewing approach is needed.
Due to classic swiss style tournaments are meant to sort out for ko semifinals the behaviour is acceptable for now.